requirements.txt
appdirs==1.4.3
click==6.7
coreapi==2.3.0
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.10.6
djangorestframework==3.6.2
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
